+++ crates/tycho-simulation/src/evm/protocol/erc4626/vm.rs
@@ -142,6 +142,7 @@ where
         overrides,
         gas_limit: None,
         transient_storage: None,
+        block_overrides: None,
     };
 
     let res = vm_engine